Phoenix ChitChat > Get tests (mostly) working for Part 2
Needed to update the tests quite a bit to get them working after all
the changes to the User<->Credential relation. Mostly working now.
There are still two assertion failures to do with
Credential.password_hash but ima leave that for now because I figure
it'll all change in Part 3 once we have actual encryption and auth.